Extract mapBriefToAmericanStenoKeys
