+++ src/components/StrokeTip.jsx
@@ -9,12 +9,13 @@ import PalantypeDiagram from "./../StenoLayout/PalantypeDiagram";
 import {
   shouldShowStroke,
   splitBriefsIntoStrokes,
-  mapBriefToAmericanStenoKeys,
   mapBriefToBrazilianPortugueseStenoKeys,
   mapBriefToDanishStenoKeys,
   mapBriefToKoreanModernCStenoKeys,
   mapBriefToPalantypeKeys,
 } from "./../utils/typey-type";
+
+import mapBriefToAmericanStenoKeys from '../utils/stenoLayouts/mapBriefToItalianMichelaStenoKeys';
 import mapBriefToItalianMichelaStenoKeys from '../utils/stenoLayouts/mapBriefToItalianMichelaStenoKeys';
 import mapBriefToJapaneseStenoKeys from '../utils/stenoLayouts/mapBriefToJapaneseStenoKeys';
